requirements.txt
Flask==2.1.1
PyPDF2==1.26.0
Pillow==8.4.0
docx2pdf==0.1.5
pdf2image==1.16.0
werkzeug==2.1.1
